pyproject.toml
[tool.poetry]
name = "chordgen"
version = "0.1.0"
description = "AI-Powered Chord Progression Generator with MIDI output and effects suggestions"
authors = ["Dante Castro <dante@balanstar.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "^1.29.0"
librosa = "^0.10.1"
soundfile = "^0.12.1"
mido = "^1.3.0"
python-rtmidi = "^1.5.8"
pyyaml = "^6.0.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
yt-dlp = "^2023.12.30"
numpy = "^1.24.0"
scipy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
